requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Async & Background Tasks
celery==5.3.4
redis==5.0.1

# PII Detection & Redaction
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
spacy==3.7.2

# OCR & Image Processing
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3

# PDF Processing
PyMuPDF==1.23.8

# Cloud Services
google-cloud-vision==3.4.5
boto3==1.29.7
botocore==1.32.7

# Integrations
jira==3.5.2
requests==2.31.0
openai==1.3.5

# NLP Model (download separately: python -m spacy download en_core_web_lg)
# en-core-web-lg==3.7.1

# Utilities
python-dotenv==1.0.0
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Production
gunicorn==21.2.0
sentry-sdk[fastapi]==1.39.1
